create new target/os "win32s". strictly speaking at this point win32s is a clone of the "nt386" target, but will allow some #ifdef blocks in the C runtime so the win32s version will be compatible with Windows 3.1 win32s. The C runtime (win32s and nt targets) does not work properly with Windows 3.1 win32s, though it works properly with Windows 95 and later, and NT 3.x and later. As far as I can tell it's a combination of a NULL pointer dereference that only happens in Windows 3.1, and the use of some Widechar (*W) functions that don't exist in Windows 3.1.
The modification is large so that all parts of the clib have a makefile
that corresponds to the new win32s target, and modifies the *.mif files
to specify when to compile win32s. It also adds a define to set WINVER
to a value appropriate for NT 3.x/Windows 3.1 for #if block use. I
apologize if this bloats the compiled binary package size but I would
rather see two different clibs for win32s vs NT than saddle the NT builds
with legacy cruft in the runtime just to support Windows 3.1.