I have achieved to get the permission from Terry Colligan at Tenberry, to redistribute latest dos4gw 2.01 into OpenWatcom, as an upgrade of current 1.97: http://www.tenberry.com/dos4g/watcom/rn4gw.html
It was a paid upgrade, that he is kindly providing to the project, on a royalte free basis.

Unfortunately his backup server got broken, and he will need about a month to recover it. In the meantime, this is upgrading dos4gw.exe to latest 2.01a, while rest of package will come later.