Fix VIW exiting when pressing weird/system keys like the Windows key
When pressing the Windows key, the ConvertWierdCharacter returns VI_KEY(
NULL ) which in one of its two uses slips through MapVirtualKeyToVIKey
down to a check added a year ago in key.c that causes the editor to
exit. This fix changes ConvertWierdCharacter to return DUMMY like the
rest of the code and fixes the issue.