File comments for some files with no introductory comment that are used by wv.
I'm going to take a crack at debugging the debugger.  It's a mess.  For example, the Locals window is totally empty when running wdw on wdw.  (When running wdw on IPDOS executables, the window lists the local variables, although the values of many are "????".)  As I begin to study the code, I will add file comments for files that don't have anything at all and that are linked into wdw.

In adding comments, I'm being conservative.  (It's better to say nothing than to add a comment that is inaccurate.)