revised file comments for some files used by wv
Most of these files contain functions that are duplicated or pseudoduplicated (same name, similar function) in multiple source files.  For such file sets, it is useful to know which flavor is used to build which executables, so I am introducing that information into the file comment.

Having that information in the file header will make it easier for me to debug the debugger, which is very buggy when run on Windows 10.