cc: cscan: Remove odd preprocessing 'extension'
Based on the comments in the code, it seems that this was hacked into
cc at some point as a feature request. However, the given behavior
wouldn't likely be accepted by other C preprocessors.

Unless there's a real case for having this hack, I'd say it's better to
clean this up.