ChangeList: The linker now warns if the number of IOPL bytes specified through the EXPORT directive is odd. The executable format and the CPU only work with words, not bytes.
Implementation note: This is an attempt to clear up confusion in the
linker source code. The linker accepts IOPL bytes, but the NE and LX
format only take words. The linker kept the number of bytes in a variable
named iopl_words but then wrote out that number such that the executable
did contain the right number of IOPL words, but might set a reserved bit
if the byte count was odd. Now iopl_words actually contains words and
odd byte counts on input are flagged. Similarly IOPL_WORD_SHIFT was defined
as 2 instead of 3 and shifted bytes rather than words; this was also
corrected.

[git-p4: depot-paths = "//depot/openwatcom/": change = 38216]