Leave room past the flush threshold in codebuf
AsmByte() writes to this buffer and calls FlushCurrSeg() when BufSize
reaches the threshold.  However, that function does not always
empty the buffer; sometimes it wants to wait for a few more bytes to
come.  If it wants to wait and the buffer is full, this would lead
to an out-of-bounds write.