Fix memory errors in wasm (#916)
* Use strcmp to compare names in AddFlist()

memcmp was used with the length of the needle string, which could
result in an out-of-bounds read if the other string was shorter.

* Don't read before start of line in replace_parm()

* Leave room past the flush threshold in codebuf

AsmByte() writes to this buffer and calls FlushCurrSeg() when BufSize
reaches the threshold.  However, that function does not always
empty the buffer; sometimes it wants to wait for a few more bytes to
come.  If it wants to wait and the buffer is full, this would lead
to an out-of-bounds write.