cmac1: Fix buffer overflow during macro expansion
As I figured out while compiling MicroPython with Open Watcom, the
comment here "size of Buffer is OK" is a lie, as I noticed wcc
segfault.

Unfortunately, I wasn't able to derive a minimal example
which replicates it, but here are steps you can use to reproduce the
bug using the same source I was (this assumes you have WATCOM and
related environment variables set correctly):

    $ git clone -b dos https://github.com/jackrosenthal/micropython
    $ cd micropython/ports/dos
    $ wcc -bt==dos -za99 -I. -I../.. -Ibuild \
          -DMICROPY_ROM_TEXT_COMPRESSION=1 -DNO_QSTR -pl -0 -zq \
          -fr ../../py/objdict.c
    ~> Segmentation fault

Using GDB and some hardware watchpoints, I was able to deduce the
segfault was caused by the scanbuf (a.k.a., the "Buffer" global
variable) overflowing and corrupting the following MCB, and it was
being happening at the code I modified in this commit.

This changes that code to instead using WriteBufferChar and
WriteBufferNullChar, which will bounds check Buffer and re-alloc when
required.