URL: Replace dead url to help users to find the recent landing Page.
The old Website looks dead since several Years.
A landing page was restored in Jan.2023 with a download links
for the old release of OW 1.9, but nothing more happens since then.

There are more locations in the git to the dead URL,
(windows resource files, example sources, documentation),
but that need discussion first.

--
Regards ... Detlef