owcc: Accept pthread and pedantic as alias for mthreads and Wextra
Most unix based Open Source Projects expect gcc or clang as compiler.
(or the gcc port mingw on windows)

OpenWatcom is old and the userbase is so small compared to gcc or clang,
that porting the affected projects to support OpenWatcom is nearly impossible.

Sometimes the only Problem when using owcc instead of gcc/clang
is a commandline option, which is not handled by owcc.

By making owcc commandline handling less strict,
OpenWatcom can be used more often as a dropin for gcc or clang.

This patch allows owcc to accept "-pthread" and "-pedantic":

-pthread option in gcc / clang:
* defines _REENTRANT (obsolete, was selecting multithread aware implementations)
* links with libpthread

With this patch, "-pthread" is an alias for the windows based option "-mthreads"
For linking, no adjustment is needed, as the
OpenWatcom runtime library already includes the pthread code.

-pedantic option in gcc / clang
* Use more checks as with "-Wall" and "-Wextra"

With this patch, "-pedantic" is an alias for "-Wextra".

--
Regards ... Detlef