clib: fix NULL pointer dereferences (#1476)
If /etc/hosts or /etc/services is missing, the corresponding
functions gethostent() and getservent() crash.
This patch avoids the problem by ading the missed NULL checks.

Note: the patch is trivial, but it was untested on the main OW
source tree. The testing was only done on a local rip-off.