Dns (#1498)
* clib: add NULL checks to getnetent() and getprotoent()

Should not crash if /etc/networks or /etc/protocols are missing.

* resolve: avoid exposing freed data in gethostbyname()

The "answer" fields that are exposed to user, are NULL-ified to
avoid them being freed at `dns_cleanup`.
Add a missing NULL-ification.

This produces a leak, since gethostbyname() is supposed to use
static storage. But leak is a separate story, and its still better
than exposing freed areas.